pyproject.toml
[tool.poetry]
name = "product-management-system"
version = "0.1.0"
description = "Product Management System API using FastAPI and MongoDB"
authors = ["el-collins <charlescollins476@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.2"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
pytest-asyncio = "^0.24.0"
pytest = "^8.3.3"
httpx = "^0.27.2"
motor = "^3.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest]
pythonpath = "."
testpaths = ["tests"]
addopts = "-v"
